refactor radard
